Move seed identification functions into dedicated package
